refactor(prosto_derive): eliminate code duplication and optimize macro generation
Major performance and maintainability improvements:

1. **Eliminated massive code duplication** (~180 lines):
   - Extracted duplicated ProtoWire delegating impl generation into
     `generate_delegating_proto_wire_impl()` shared function
   - Extracted duplicated ProtoExt sun impl generation into
     `generate_sun_proto_ext_impl()` shared function
   - Removed identical code blocks from structs.rs, enums.rs, and complex_enums.rs

2. **Performance improvements**:
   - Reduced macro expansion overhead by centralizing code generation
   - Cleaner generated code should enable better compiler optimizations
   - Eliminated redundant TokenStream constructions during macro expansion

3. **Code quality improvements**:
   - Single source of truth for delegating implementations
   - Easier to maintain and update implementation logic
   - Reduced risk of divergence between struct/enum/complex_enum implementations

These changes improve compile-time performance for crates using prosto_derive
and make future maintenance significantly easier.
