+++ crates/prosto_derive/src/proto_message/unified_field_handler.rs
@@ -510,6 +510,98 @@ pub fn build_encode_stmts(fields: &[FieldInfo<'_>], base: &TokenStream2) -> Vec<
         .collect()
 }
 
+/// Generate delegating ProtoWire implementation for a sun type
+/// This eliminates code duplication across structs, enums, and complex enums
+pub fn generate_delegating_proto_wire_impl(
+    shadow_ty: &TokenStream2,
+    target_ty: &syn::Type,
+) -> TokenStream2 {
+    quote! {
+        impl ::proto_rs::ProtoWire for #target_ty {
+            type EncodeInput<'a> = <#shadow_ty as ::proto_rs::ProtoShadow<#target_ty>>::Sun<'a>;
+            const KIND: ::proto_rs::ProtoKind = <#shadow_ty as ::proto_rs::ProtoWire>::KIND;
+
+            #[inline(always)]
+            fn proto_default() -> Self {
+                <#shadow_ty as ::proto_rs::ProtoShadow<#target_ty>>::to_sun(
+                    <#shadow_ty as ::proto_rs::ProtoWire>::proto_default(),
+                )
+                .expect("default shadow must be decodable")
+            }
+
+            #[inline(always)]
+            fn clear(&mut self) {
+                *self = Self::proto_default();
+            }
+
+            #[inline(always)]
+            fn is_default_impl(value: &Self::EncodeInput<'_>) -> bool {
+                let shadow = <#shadow_ty as ::proto_rs::ProtoShadow<#target_ty>>::from_sun(*value);
+                <#shadow_ty as ::proto_rs::ProtoWire>::is_default_impl(&shadow)
+            }
+
+            #[inline(always)]
+            unsafe fn encoded_len_impl_raw(value: &Self::EncodeInput<'_>) -> usize {
+                let shadow = <#shadow_ty as ::proto_rs::ProtoShadow<#target_ty>>::from_sun(*value);
+                <#shadow_ty as ::proto_rs::ProtoWire>::encoded_len_impl_raw(&shadow)
+            }
+
+            #[inline(always)]
+            fn encode_raw_unchecked(
+                value: Self::EncodeInput<'_>,
+                buf: &mut impl ::proto_rs::bytes::BufMut,
+            ) {
+                let shadow = <#shadow_ty as ::proto_rs::ProtoShadow<#target_ty>>::from_sun(value);
+                <#shadow_ty as ::proto_rs::ProtoWire>::encode_raw_unchecked(shadow, buf)
+            }
+
+            #[inline(always)]
+            fn decode_into(
+                wire_type: ::proto_rs::encoding::WireType,
+                value: &mut Self,
+                buf: &mut impl ::proto_rs::bytes::Buf,
+                ctx: ::proto_rs::encoding::DecodeContext,
+            ) -> Result<(), ::proto_rs::DecodeError> {
+                let mut shadow = <#shadow_ty as ::proto_rs::ProtoWire>::proto_default();
+                <#shadow_ty as ::proto_rs::ProtoWire>::decode_into(wire_type, &mut shadow, buf, ctx)?;
+                *value = <#shadow_ty as ::proto_rs::ProtoShadow<#target_ty>>::to_sun(shadow)?;
+                Ok(())
+            }
+        }
+    }
+}
+
+/// Generate sun-based ProtoExt implementation
+/// This eliminates code duplication across different type handlers
+pub fn generate_sun_proto_ext_impl(
+    shadow_ty: &TokenStream2,
+    target_ty: &syn::Type,
+    decode_arms: &[TokenStream2],
+    post_decode_impl: &TokenStream2,
+) -> TokenStream2 {
+    quote! {
+        impl ::proto_rs::ProtoExt for #target_ty {
+            type Shadow<'b> = #shadow_ty where Self: 'b;
+
+            #[inline(always)]
+            fn merge_field(
+                value: &mut Self::Shadow<'_>,
+                tag: u32,
+                wire_type: ::proto_rs::encoding::WireType,
+                buf: &mut impl ::proto_rs::bytes::Buf,
+                ctx: ::proto_rs::encoding::DecodeContext,
+            ) -> Result<(), ::proto_rs::DecodeError> {
+                match tag {
+                    #(#decode_arms,)*
+                    _ => ::proto_rs::encoding::skip_field(wire_type, tag, buf, ctx),
+                }
+            }
+
+            #post_decode_impl
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
